Optimise pulse capturing
Each cycle of capturing a pulse was taking between 30 and 60us, which is
too slow to capture pulses that are either below or above 50us. By
making changes to how `_capture_pulses` is implemented. Main changes
are:

 * Cache pin locally
 * Use preallocated bytearray for puless
 * Use the MicroPython native code emitter

During tests base tests, checking the sensor every second for 3 minutes,
brought the error rate down from 5-8% to 0%.